+++ src/lib.rs
@@ -3,10 +3,12 @@
 /// The [`Middleware`] trait and some common middleware primitives.
 pub mod middleware;
 
+#[cfg(feature = "jwt")]
+pub mod jwt;
 /// Synchronous API for Feather.
 mod sync;
 
-pub use crate::middleware::{Middleware,MiddlewareResult};
-pub use feather_runtime::http::HttpRequest as Request;
-pub use feather_runtime::http::HttpResponse as Response;
+pub use feather_runtime::http::Request;
+pub use feather_runtime::http::Response;
+pub use middleware::MiddlewareResult;
 pub use sync::App;