+++ src/middleware/builtins.rs
@@ -1,28 +1,10 @@
-use std::fs;
-use std::fs::File;
-use std::io::Read;
-use std::path::Path;
-use dyn_clone::DynClone;
-use feather_runtime::http::HttpRequest as Request;
-use feather_runtime::http::HttpResponse as Response;
-/// Common trait for all middleware types. Implemented automatically for functions fitting
-/// the `(request, response) -> result` signature.
-
-pub trait Middleware: Send + Sync + DynClone {
-    /// Handle an incoming request by transforming it into a response.
-    fn handle(&self, request: &mut Request, response: &mut Response) -> MiddlewareResult;
-}
-dyn_clone::clone_trait_object!(Middleware);
-
-/// MiddlewareResult is used to control the flow of middleware execution.
-/// 
-/// It can be used to skip all subsequent middleware and continue to the next route.
-pub enum MiddlewareResult {
-    /// Continue to the next middleware.
-    Next,
-    /// Skip all subsequent middleware and continue to the next route.
-    NextRoute,
-}
+use super::common::{Middleware, MiddlewareResult};
+use feather_runtime::http::{Request, Response};
+use std::{
+    fs::{self, File},
+    io::Read,
+    path::Path,
+};
 
 #[derive(Clone)]
 /// Log incoming requests and transparently pass them to the next middleware.
@@ -63,7 +45,7 @@ impl Middleware for Cors {
 pub struct ServeStatic(String);
 
 impl ServeStatic {
-    #[must_use= "Put this in a `App.use_middleware()` Method"]
+    #[must_use = "Put this in a `App.use_middleware()` Method"]
     pub const fn new(directory: String) -> Self {
         Self(directory)
     }
@@ -72,75 +54,50 @@ impl ServeStatic {
 impl Middleware for ServeStatic {
     fn handle(&self, request: &mut Request, response: &mut Response) -> MiddlewareResult {
         let wanted_path = request.uri.to_string();
-        let dir = fs::read_dir(Path::new(self.0.as_str())).expect(format!("Error While Reading the {}",self.0).as_str());
-        dir.for_each(|entry|{
+        let dir = fs::read_dir(Path::new(self.0.as_str()))
+            .expect(format!("Error While Reading the {}", self.0).as_str());
+        dir.for_each(|entry| {
             let entry = entry.unwrap();
             let enter_path = entry.path();
             let mut file = File::open(entry.path()).unwrap();
             let ext = enter_path.extension().unwrap();
             match ext.to_str().unwrap() {
                 "png" => {
-                    if wanted_path.contains(entry.file_name().into_string().unwrap().as_str()){
+                    if wanted_path.contains(entry.file_name().into_string().unwrap().as_str()) {
                         response.add_header("Content-Type", "image/png");
                         let mut buf = Vec::with_capacity(4096);
                         file.read(&mut buf).unwrap();
                         response.send_bytes(buf);
                     }
                 }
-                "jpg" =>{
-                    if wanted_path.contains(entry.file_name().into_string().unwrap().as_str()){
+                "jpg" => {
+                    if wanted_path.contains(entry.file_name().into_string().unwrap().as_str()) {
                         response.add_header("Content-Type", "image/jpg");
                         let mut buf = Vec::with_capacity(1024);
                         file.read_to_end(&mut buf).unwrap();
                         response.send_bytes(buf);
                     }
                 }
-                "jpeg" =>{
-                    if wanted_path.contains(entry.file_name().into_string().unwrap().as_str()){
+                "jpeg" => {
+                    if wanted_path.contains(entry.file_name().into_string().unwrap().as_str()) {
                         response.add_header("Content-Type", "image/jpeg");
                         let mut buf = Vec::with_capacity(1024);
                         file.read_to_end(&mut buf).unwrap();
                         response.send_bytes(buf);
                     }
                 }
-                "html" =>{
-                    if wanted_path.contains(entry.file_name().into_string().unwrap().as_str()){
+                "html" => {
+                    if wanted_path.contains(entry.file_name().into_string().unwrap().as_str()) {
                         response.add_header("Content-Type", "text/html");
                         let mut buf = Vec::with_capacity(1024);
                         file.read_to_end(&mut buf).unwrap();
                         response.send_bytes(buf);
                     }
                 }
-                
-                
+
                 _ => unreachable!(),
             }
-            
         });
         MiddlewareResult::Next
-    }
-}
-
-/// Implement the `Middleware` trait for a slice of middleware.
-impl Middleware for [&Box<dyn Middleware>] {
-    fn handle(&self, request: &mut Request, response: &mut Response) -> MiddlewareResult {
-        for middleware in self {
-            if matches!(
-                middleware.handle(request, response),
-                MiddlewareResult::NextRoute
-            ) {
-                return MiddlewareResult::NextRoute;
-            }
-        }
-        MiddlewareResult::Next
-    }
-}
-
-///Implement the `Middleware` trait for Closures with Request and Response Parameters.
-impl<F: Fn(&mut Request, &mut Response) -> MiddlewareResult + Sync + Send + DynClone> Middleware
-    for F
-{
-    fn handle(&self, request: &mut Request, response: &mut Response) -> MiddlewareResult {
-        self(request, response)
     }
 }